[fix] 게시글 작성 지역 추가 (#130)
* [feat] 유기동물 공고 자동 업데이트 기능 구현

* [fix] 유기동물 공조 조회 response 변경 (#125)

* [fix] 게시글 삭제 시 댓글 존재 여부 확인 추가 (#128)

* [fix] 게시글 작성 지역 추가

+++ src/main/java/challengers/findog/src/board/BoardService.java
@@ -31,6 +31,7 @@ public class BoardService {
             int postId;
             String title = postBoardReq.getTitle();
             int category = postBoardReq.getCategory();
+            int region = postBoardReq.getRegion();
             String content = postBoardReq.getContent();
             ArrayList<MultipartFile> imgFiles = postBoardReq.getImgFiles();
             if (title.isEmpty() || title.isBlank()) {
@@ -38,6 +39,9 @@ public class BoardService {
             }
             if (category == 0 || category > 4) {
                 throw new BaseException(EMPTY_CATEGORY);
+            }
+            if (region == 0 || region > 13) {
+                throw new BaseException(EMPTY_REGION);
             }
             if (content.isEmpty() || content.isBlank()) {
                 throw new BaseException(EMPTY_CONTENT);
@@ -107,9 +111,10 @@ public class BoardService {
     //게시글 삭제, 해당 게시글 댓글 함께 삭제
     @Transactional(rollbackFor = Exception.class)
     public void deleteBoard(int postId) throws BaseException {
-        //todo 댓글 삭제
-        if(boardRepository.deleteCommentByPostId(postId) == 0) {
-            throw new BaseException(FAIL_DELETE_COMMENTS);
+        if(boardRepository.existComment(postId) == 1) {
+            if (boardRepository.deleteCommentByPostId(postId) == 0) {
+                throw new BaseException(FAIL_DELETE_COMMENTS);
+            }
         }
         //이미지 삭제
         List<String> arr = boardRepository.checkImg(postId); //기존 저장된 img 유무 확인
